+++ CHANGES.rst
@@ -4,6 +4,17 @@ Changelog
 0.7dev
 ------
 
+- Extend ``CartItemCalculator`` by ``item_net``, ``item_vat`` and
+  ``item_weight`` functions.
+  [rnix]
+
+- Deprecate ``FlatRate`` shipping.
+  [rnix]
+
+- Add ``DefaultShipping`` and corresponding control panel settings. GS upgrade
+  2_to_3 must be applied.
+  [rnix]
+
 - Fix displaying of discounted price. Original price only gets displayed if
   it deferrs from discounted one.
   [rnix]